fix: refactor kafka tests, they were actually integration tests
